requirements.txt
asgiref==3.9.1
Django==5.2.4
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.1
